Add gcp.go from client-go to cloud-provider-gcp as that file is being deprecated from client-go.
If users do not want to use an additional binary (gke-gcloud-auth-plugin),
then we they can reference and build this library in to their custom k8s
clients to provide the same functionality. This use case becomes
necessary if customers are unwilling to add additional binaries on their
system to avoid remote code execution attacks.